+++ CHANGELOG.md
@@ -19,6 +19,7 @@ options. The project adheres to semantic versioning.
 * Fix a warning about possible loss of data.
 * Fix version numbering in CMakeLists.txt
 * Remove unused declaration of the undefined `ParseResult::get_option`.
+* Throw on invalid option syntax when beginning with a `-`.
 
 ## 2.1.1